requirements.txt
alembic==1.11.1
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
Flask==2.3.2
Flask-API==3.1
Flask-Migrate==4.0.4
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
